requirements.txt
Django==4.1.5
django-webpack-loader==1.8.0
djangorestframework==3.14.0
numpy==1.23.4
pandas==1.5.0
scikit-learn==1.1.3
xgboost==1.7.3
psycopg2>=2.8
